Abstract out elements
There's a bug in the scroll top and the navigate back to help page on the 0th section still.
