+++ src/screens/help/HelpAddSighting.jsx
@@ -1,57 +1,25 @@
-import React, { useState } from 'react';
-import {
-  TouchableOpacity,
-  Text,
-  Image,
-  ScrollView,
-  View,
-  StyleSheet,
-} from 'react-native';
+import React, { useRef, useState } from 'react';
+import { Image, ScrollView, View, StyleSheet } from 'react-native';
 import screens from '../../constants/screens';
 import theme from '../../constants/theme';
-import globalStyles from '../../styles/globalStyles';
-import Typography from '../../components/Typography';
 import { Dimensions } from 'react-native';
 import newSighting1 from '../../../assets/images/HelpPage/NewSighting1.png';
 import newSighting2_1 from '../../../assets/images/HelpPage/NewSighting2.1.png';
 import newSighting2_2 from '../../../assets/images/HelpPage/NewSighting2.2.png';
 import newSighting3 from '../../../assets/images/HelpPage/NewSighting3.png';
 import newSighting4 from '../../../assets/images/HelpPage/NewSighting4.png';
-
-const winWidth = Dimensions.get('window').width;
-const imageRatio = winWidth / 700; // The 700 is the px width of the images
-
-const Instruction = (props) => {
-  const typographyID = `HELP_ADD_SIGHTING_${props.number}`;
-  return (
-    <View style={styles.instruction}>
-      <Text
-        style={[
-          globalStyles.basicText,
-          styles.instructionSubElement,
-          styles.instructionNumber,
-        ]}
-      >
-        {props.number}
-      </Text>
-      <Typography
-        id={typographyID}
-        style={[
-          globalStyles.basicText,
-          styles.instructionSubElement,
-          styles.instructionText,
-        ]}
-      />
-    </View>
-  );
-};
+import { Instruction, PageNavigationButtons } from './HelpElements';
 
 const HelpAddSighting = ({ navigation }) => {
   const [helpSection, setHelpSection] = useState(0);
+  const scrollReference = useRef();
+  const scrollTop = () => {
+    scrollReference.current?.scrollTo({ y: 0 });
+  };
 
   return (
     <View style={styles.content}>
-      <ScrollView style={styles.helpPage}>
+      <ScrollView style={styles.helpPage} ref={scrollReference}>
         {helpSection === 0 && (
           <>
             <Image
@@ -59,22 +27,12 @@ const HelpAddSighting = ({ navigation }) => {
               source={newSighting1}
               resizeMode={'contain'}
             />
-            <Instruction number={1} />
-            <View style={[styles.horizontal, styles.bottomElement]}>
-              {/* // TODO: why not working?? */}
-              <TouchableOpacity
-                onPress={() => navigation.navigate(screens.helpPage)}
-              >
-                <View style={[styles.button, styles.buttonInactive]}>
-                  <Typography id="BACK" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-              <TouchableOpacity onPress={() => setHelpSection(1)}>
-                <View style={styles.button}>
-                  <Typography id="NEXT" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-            </View>
+            <Instruction section={'HELP_ADD_SIGHTING'} number={1} />
+            <PageNavigationButtons
+              onBackward={() => navigation.navigate(screens.helpPage)}
+              onForward={() => setHelpSection(1)}
+              postNavigate={scrollTop}
+            />
           </>
         )}
         {helpSection === 1 && (
@@ -84,20 +42,13 @@ const HelpAddSighting = ({ navigation }) => {
               source={newSighting2_1}
               resizeMode={'contain'}
             />
-            <Instruction number={2} />
-            <Instruction number={3} />
-            <View style={[styles.horizontal, styles.bottomElement]}>
-              <TouchableOpacity onPress={() => setHelpSection(0)}>
-                <View style={[styles.button, styles.buttonInactive]}>
-                  <Typography id="BACK" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-              <TouchableOpacity onPress={() => setHelpSection(2)}>
-                <View style={styles.button}>
-                  <Typography id="NEXT" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-            </View>
+            <Instruction section={'HELP_ADD_SIGHTING'} number={2} />
+            <Instruction section={'HELP_ADD_SIGHTING'} number={3} />
+            <PageNavigationButtons
+              onBackward={() => setHelpSection(0)}
+              onForward={() => setHelpSection(2)}
+              postNavigate={scrollTop}
+            />
           </>
         )}
         {helpSection === 2 && (
@@ -107,20 +58,13 @@ const HelpAddSighting = ({ navigation }) => {
               source={newSighting2_2}
               resizeMode={'contain'}
             />
-            <Instruction number={4} />
-            <Instruction number={5} />
-            <View style={[styles.horizontal, styles.bottomElement]}>
-              <TouchableOpacity onPress={() => setHelpSection(1)}>
-                <View style={[styles.button, styles.buttonInactive]}>
-                  <Typography id="BACK" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-              <TouchableOpacity onPress={() => setHelpSection(3)}>
-                <View style={styles.button}>
-                  <Typography id="NEXT" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-            </View>
+            <Instruction section={'HELP_ADD_SIGHTING'} number={4} />
+            <Instruction section={'HELP_ADD_SIGHTING'} number={5} />
+            <PageNavigationButtons
+              onBackward={() => setHelpSection(1)}
+              onForward={() => setHelpSection(3)}
+              postNavigate={scrollTop}
+            />
           </>
         )}
         {helpSection === 3 && (
@@ -130,20 +74,13 @@ const HelpAddSighting = ({ navigation }) => {
               source={newSighting3}
               resizeMode={'contain'}
             />
-            <Instruction number={6} />
-            <Instruction number={7} />
-            <View style={[styles.horizontal, styles.bottomElement]}>
-              <TouchableOpacity onPress={() => setHelpSection(2)}>
-                <View style={[styles.button, styles.buttonInactive]}>
-                  <Typography id="BACK" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-              <TouchableOpacity onPress={() => setHelpSection(4)}>
-                <View style={styles.button}>
-                  <Typography id="NEXT" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-            </View>
+            <Instruction section={'HELP_ADD_SIGHTING'} number={6} />
+            <Instruction section={'HELP_ADD_SIGHTING'} number={7} />
+            <PageNavigationButtons
+              onBackward={() => setHelpSection(2)}
+              onForward={() => setHelpSection(4)}
+              postNavigate={scrollTop}
+            />
           </>
         )}
         {helpSection === 4 && (
@@ -153,30 +90,25 @@ const HelpAddSighting = ({ navigation }) => {
               source={newSighting4}
               resizeMode={'contain'}
             />
-            <Instruction number={8} />
-            <View style={[styles.horizontal, styles.bottomElement]}>
-              <TouchableOpacity onPress={() => setHelpSection(3)}>
-                <View style={[styles.button, styles.buttonInactive]}>
-                  <Typography id="BACK" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-              <TouchableOpacity
-                onPress={() => {
-                  navigation.navigate(screens.home);
-                  setHelpSection(0);
-                }}
-              >
-                <View style={styles.button}>
-                  <Typography id="DONE" style={globalStyles.buttonText} />
-                </View>
-              </TouchableOpacity>
-            </View>
+            <Instruction section={'HELP_ADD_SIGHTING'} number={8} />
+            <PageNavigationButtons
+              onBackward={() => setHelpSection(3)}
+              onForward={() => {
+                navigation.navigate(screens.home);
+                setHelpSection(0);
+              }}
+              forwardName={'DONE'}
+              postNavigate={scrollTop}
+            />
           </>
         )}
       </ScrollView>
     </View>
   );
 };
+
+const winWidth = Dimensions.get('window').width;
+const imageRatio = winWidth / 700; // The 700 is the px width of the images
 
 const styles = StyleSheet.create({
   content: {
@@ -191,53 +123,6 @@ const styles = StyleSheet.create({
   helpImage: {
     width: winWidth,
     height: 800 * imageRatio,
-  },
-  instruction: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginHorizontal: 15,
-    marginVertical: 5,
-    borderBottomColor: '#2C2C2C',
-    borderBottomWidth: 0.5,
-  },
-  instructionSubElement: {
-    flex: 0,
-    marginBottom: 10,
-    marginHorizontal: 10,
-    paddingVertical: 5,
-    textAlignVertical: 'center',
-    lineHeight: 25,
-    color: theme.black,
-  },
-  instructionNumber: {
-    paddingTop: 18,
-    fontSize: 30,
-    fontWeight: 'bold',
-    color: theme.green,
-  },
-  instructionText: {
-    paddingRight: 30,
-    flexGrow: 1,
-  },
-  button: {
-    backgroundColor: theme.green,
-    paddingVertical: 12,
-    paddingHorizontal: 50,
-    borderRadius: 20,
-    margin: '5%',
-  },
-  horizontal: {
-    width: '100%',
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-evenly',
-  },
-  buttonInactive: {
-    backgroundColor: '#CACACA',
-  },
-  bottomElement: {
-    marginTop: 7,
-    marginBottom: 15,
   },
 });
